+++ Changes.md
@@ -11,6 +11,7 @@ Version 3.3.0.6
   8. Fix section name box does not fill image container when inside with the Boost theme.
   9. Put back 'numsections'.  Thanks to "King's College London" for funding this enhancement.
  10. Fix existing courses have an incorrect number of courses - #90.
+ 11. Fix "Editing section summary and changing summary text breaks grid image" - #92.
 
 Version 3.3.0.5
   1. Fix incorrect help for 'sectiontitleboxinsideposition' setting.